Add more rest apis for the newly created BookService. Add a resource not found exception as well a way to mark book as owned since the plan is to have all the book of a series in thelibrary but you may not  own all of them. So a new field called owned was added.

+++ src/main/java/com/mangabooks/library/Entity/BookSeries.java
@@ -32,7 +32,7 @@ public class BookSeries {
     private List<Author> authors;
 
     @Column(name ="total_volumes")
-    private Integer totalVolumes;
+    private Byte totalVolumes;
 
 
 }
